backend/requirements-dev.txt
-r requirements.txt
ipython==8.31.0
pre-commit==4.1.0
pylint-django==2.6.1
pylint==3.3.3
pytest-django==4.9.0
pytest==8.3.4
python-dotenv==1.0.1
requirementscheck==0.0.5
types-requests==2.32.0.20241016
